refactor: split PlaylistSelection styled components into grouped modules
