+++ src/components/PlaylistSelection/styled.layout.ts
@@ -1,0 +1,76 @@
+import styled from 'styled-components';
+import { Card } from '../styled';
+import { theme } from '@/styles/theme';
+
+export const Container = styled.div<{ $inDrawer?: boolean }>`
+  width: 100%;
+  display: flex;
+  ${({ $inDrawer }) =>
+    $inDrawer
+      ? `
+    flex: 1;
+    min-height: 0;
+    flex-direction: column;
+    align-items: stretch;
+    justify-content: flex-start;
+    padding: 0 ${theme.spacing.lg} ${theme.spacing.md};
+    box-sizing: border-box;
+  `
+      : `
+    min-height: 100vh;
+    min-height: 100dvh;
+    align-items: center;
+    justify-content: center;
+    padding: 1rem;
+
+    @media (max-width: ${theme.breakpoints.lg}) {
+      padding: 0.5rem;
+    }
+  `}
+`;
+
+export const SelectionCard = styled(Card)<{ $maxWidth: number; $inDrawer?: boolean }>`
+  width: 100%;
+  max-width: ${({ $maxWidth, $inDrawer }) => ($inDrawer ? 'none' : `${$maxWidth}px`)};
+  ${({ $inDrawer }) =>
+    $inDrawer
+      ? `
+    flex: 1;
+    min-height: 0;
+    display: flex;
+    flex-direction: column;
+    overflow: hidden;
+    background: none;
+    background-color: transparent;
+    backdrop-filter: none;
+    border: none;
+    border-radius: 0;
+    box-shadow: none;
+    padding: 0;
+    margin: 0;
+  `
+      : `
+    background: ${theme.colors.muted.background};
+    backdrop-filter: blur(12px);
+    border: 1px solid ${theme.colors.control.border};
+    border-radius: 1.25rem;
+    box-shadow: ${theme.shadows.albumArt};
+    display: flex;
+    flex-direction: column;
+    max-height: min(90dvh, 900px);
+  `}
+`;
+
+export const DrawerContentWrapper = styled.div`
+  flex: 1;
+  min-height: 0;
+  display: flex;
+  flex-direction: column;
+  padding: 0 ${theme.spacing.md};
+  overflow: hidden;
+  max-width: 1000px;
+  margin-left: auto;
+  margin-right: auto;
+  width: 100%;
+  box-sizing: border-box;
+`;
